Break down the db actions in more specific actions
