+++ src/renderer/actions/connections.js
@@ -1,0 +1,29 @@
+import { services } from '../../browser/remote';
+
+
+export const CONNECTION_REQUEST = 'CONNECTION_REQUEST';
+export const CONNECTION_SUCCESS = 'CONNECTION_SUCCESS';
+export const CONNECTION_FAILURE = 'CONNECTION_FAILURE';
+
+
+export function connect (serverId, database) {
+  return async (dispatch, getState) => {
+    const { servers } = getState();
+    const id = parseInt(serverId, 10);
+    const server = {
+      ...servers.items[id],
+      id,
+    };
+
+    console.info('actions:db:connect', server, database);
+    dispatch({ type: CONNECTION_REQUEST, server, database });
+    try {
+      await services.db.connect(server, database);
+      console.info('actions:db:connected');
+      dispatch({ type: CONNECTION_SUCCESS, server, database });
+    } catch (error) {
+      console.info('actions:db:error', error);
+      dispatch({ type: CONNECTION_FAILURE, server, database, error });
+    }
+  };
+}
