+++ src/renderer/actions/servers.js
@@ -1,18 +1,16 @@
 import { services } from '../../browser/remote';
 
 
-import {
-  LOAD_SERVERS_REQUEST,
-  LOAD_SERVERS_SUCCESS,
-  LOAD_SERVERS_FAILURE,
-  SAVE_SERVER_REQUEST,
-  SAVE_SERVER_SUCCESS,
-  SAVE_SERVER_FAILURE,
-  REMOVE_SERVER_REQUEST,
-  REMOVE_SERVER_SUCCESS,
-  REMOVE_SERVER_FAILURE,
-  FILTER_SERVERS,
-} from './types';
+export const LOAD_SERVERS_REQUEST = 'LOAD_SERVERS_REQUEST';
+export const LOAD_SERVERS_SUCCESS = 'LOAD_SERVERS_SUCCESS';
+export const LOAD_SERVERS_FAILURE = 'LOAD_SERVERS_FAILURE';
+export const SAVE_SERVER_REQUEST = 'SAVE_SERVER_REQUEST';
+export const SAVE_SERVER_SUCCESS = 'SAVE_SERVER_SUCCESS';
+export const SAVE_SERVER_FAILURE = 'SAVE_SERVER_FAILURE';
+export const REMOVE_SERVER_REQUEST = 'REMOVE_SERVER_REQUEST';
+export const REMOVE_SERVER_SUCCESS = 'REMOVE_SERVER_SUCCESS';
+export const REMOVE_SERVER_FAILURE = 'REMOVE_SERVER_FAILURE';
+export const FILTER_SERVERS = 'FILTER_SERVERS';
 
 
 export function loadServers() {
