--- src/renderer/actions/types.js
@@ -1,18 +1,0 @@
-export const LOAD_DATABASES_REQUEST = 'LOAD_DATABASES_REQUEST';
-export const LOAD_DATABASES_SUCCESS = 'LOAD_DATABASES_SUCCESS';
-export const LOAD_DATABASES_FAILURE = 'LOAD_DATABASES_FAILURE';
-export const FILTER_DATABASES = 'FILTER_DATABASES';
-export const FILTER_SERVERS = 'FILTER_SERVERS';
-export const LOAD_SERVERS_REQUEST = 'LOAD_SERVERS_REQUEST';
-export const LOAD_SERVERS_SUCCESS = 'LOAD_SERVERS_SUCCESS';
-export const LOAD_SERVERS_FAILURE = 'LOAD_SERVERS_FAILURE';
-export const SAVE_SERVER_REQUEST = 'SAVE_SERVER_REQUEST';
-export const SAVE_SERVER_SUCCESS = 'SAVE_SERVER_SUCCESS';
-export const SAVE_SERVER_FAILURE = 'SAVE_SERVER_FAILURE';
-export const REMOVE_SERVER_REQUEST = 'REMOVE_SERVER_REQUEST';
-export const REMOVE_SERVER_SUCCESS = 'REMOVE_SERVER_SUCCESS';
-export const REMOVE_SERVER_FAILURE = 'REMOVE_SERVER_FAILURE';
-export const EXECUTE_QUERY_REQUEST = 'EXECUTE_QUERY_REQUEST';
-export const EXECUTE_QUERY_SUCCESS = 'EXECUTE_QUERY_SUCCESS';
-export const EXECUTE_QUERY_FAILURE = 'EXECUTE_QUERY_FAILURE';
-export const UPDATE_SQL_SUCCESS = 'UPDATE_SQL_SUCCESS';
